refactor: Simplify version display and extract duplicated worktree utility
- Simplify version display by removing [NO-WT] debug marker in Sidebar
- Extract duplicated getCurrentWorktreeName function into shared utility module
- Update imports in main/src/index.ts and main/src/ipc/updater.ts

These changes improve code maintainability by following DRY principles and
cleaning up the UI display.
