+++ main/src/index.ts
@@ -19,6 +19,7 @@ import { StravuAuthManager } from './services/stravuAuthManager';
 import { StravuNotebookService } from './services/stravuNotebookService';
 import { Logger } from './utils/logger';
 import { setCrystalDirectory } from './utils/crystalDirectory';
+import { getCurrentWorktreeName } from './utils/worktreeUtils';
 import { registerIpcHandlers } from './ipc';
 import { setupAutoUpdater } from './autoUpdater';
 import { setupEventListeners } from './events';
@@ -28,23 +29,6 @@ import { setupConsoleWrapper } from './utils/consoleWrapper';
 import * as fs from 'fs';
 
 let mainWindow: BrowserWindow | null = null;
-
-/**
- * Extract worktree name from the current working directory path
- * Returns the worktree name if running in a worktree, undefined if in main repository
- */
-function getCurrentWorktreeName(cwd: string): string | undefined {
-  try {
-    // Match worktrees directory followed by worktree name
-    // Handles both Unix (/) and Windows (\) path separators
-    // For paths like "worktrees/feature/dev-mode-worktree-label", captures "feature/dev-mode-worktree-label"
-    const worktreeMatch = cwd.match(/worktrees[\/\\](.+)/);
-    return worktreeMatch ? worktreeMatch[1] : undefined;
-  } catch (error) {
-    console.log('Could not extract worktree name:', error);
-    return undefined;
-  }
-}
 
 /**
  * Set the application title based on development mode and worktree
