+++ main/src/ipc/updater.ts
@@ -4,23 +4,7 @@ import type { AppServices } from './types';
 import * as fs from 'fs';
 import * as path from 'path';
 import { commandExecutor } from '../utils/commandExecutor';
-
-/**
- * Extract worktree name from the current working directory path
- * Returns the worktree name if running in a worktree, undefined if in main repository
- */
-function getCurrentWorktreeName(cwd: string): string | undefined {
-  try {
-    // Match worktrees directory followed by worktree name
-    // Handles both Unix (/) and Windows (\) path separators
-    // For paths like "worktrees/feature/dev-mode-worktree-label", captures "feature/dev-mode-worktree-label"
-    const worktreeMatch = cwd.match(/worktrees[\/\\](.+)/);
-    return worktreeMatch ? worktreeMatch[1] : undefined;
-  } catch (error) {
-    console.log('Could not extract worktree name:', error);
-    return undefined;
-  }
-}
+import { getCurrentWorktreeName } from '../utils/worktreeUtils';
 
 export function registerUpdaterHandlers(ipcMain: IpcMain, { app, versionChecker }: AppServices): void {
   // Version checking handlers
